refactor(models): replace all interfaces with types
