--- src/app/core/models/book-response.model.ts
@@ -1,7 +1,0 @@
-export interface BookResponse {
-  id: string;
-  title: string;
-  author: string;
-  genre: string;
-  publicationYear: string;
-}
