--- src/app/core/models/book.model.ts
@@ -1,7 +1,0 @@
-export interface Book {
-  id: number;
-  title: string;
-  author: string;
-  genre: string;
-  publicationYear: Date;
-}
