+++ src/app/core/models/book.type.ts
@@ -1,0 +1,12 @@
+export type BookResponse = {
+  id: string;
+  title: string;
+  author: string;
+  genre: string;
+  publicationYear: string;
+}
+
+export type Book = Omit<BookResponse, 'id' | 'publicationYear'> & {
+  id: number;
+  publicationYear: Date;
+};
